Remove separate Table64Lowering pass
This pass is now just part of Memory64Lowering.

Once this lands, we can remove the `--table64-lowering` flag from
emscripten.  Because I've used an alias here there will be some interim
period where emscripten will run this pass twice since it passed both
flags.  However, this will only be temporary and that second run will
be a no-op since the first one will remove the feature.
