+++ src/passes/pass.cpp
@@ -273,8 +273,8 @@ void PassRegistry::registerPasses() {
                "32-bit one",
                createMemory64LoweringPass);
   registerPass("table64-lowering",
-               "lower 64-bit tables 32-bit ones",
-               createTable64LoweringPass);
+               "alias for memory64-lowering",
+               createMemory64LoweringPass);
   registerPass("llvm-memory-copy-fill-lowering",
                "Lower memory.copy and memory.fill to wasm mvp and disable "
                "the bulk-memory feature.",
